finishing dashboard

+++ Harris.Core/ViewModels/MatrixViewModel.cs
@@ -9,7 +9,7 @@ namespace Harris.Core.ViewModels {
   public class MatrixViewModel {
     public Contract Contract { get; set; }
     public double CategoryMatch { get; set; }
-    public int CPARScore { get; set; }
+    public double CPARScore { get; set; }
     public double BestMatch { get; set; }
     public string StartDate { get; set; }
     public string EndDate { get; set; }
